argonuat 6.1 full migration

+++ src/main/scala/jsentric/Validator.scala
@@ -82,8 +82,8 @@ trait Validators {
     def message(n: Number): String
 
     def maybeValid(path: Path) = {
-      case (Some(n), _) if n.isNumber && doubleFail(n.number.get) =>
-        message(n.number.get) -> path
+      case (Some(n), _) if n.isNumber && doubleFail(n.number.get.toDouble) =>
+        message(n.number.get.toDouble) -> path
     }
   }
